refactor: adjust structure and styles
