+++ src/components/App.css
@@ -1,5 +1,29 @@
+.main_container{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
 .container{
     width:75%;
+}
+.breadcrumbs_container{
+    z-index: 99;
+    line-height: 2;
+    text-align: center;
+    color: #fe1f4f;
+    margin-bottom: 10px;
+    border: 1px solid;
+    border-radius: 10px;
+    justify-content: space-around;
+    font-weight: 500;
+    background-color: #230f1b;
+    user-select: none;
+}
+.breadcrumbs_item{
+    display: block;
+    padding-left: 12px;
+    padding-right: 12px;
+    color: #fff;
 }
 .items {
     margin-right: 25px;
@@ -175,7 +199,7 @@
     box-shadow: none;
 }
 
-.rhap_main-controls-button,.rhap_volume-button {
+.rhap_main-controls-button,.rhap_volume-button, .current {
     color: #fe1f4f;
 }
